# openapi.yaml
openapi: 3.0.3
info:
  title: Homegate API
  description: |
    #### WARNING: This OpenAPI specification is AI generated. It may not be accurate.

    A backend service to gatekeep Pubky Homeserver signups. Provides four verification methods:
    - SMS verification via Prelude
    - Lightning Network payment verification
    - IP-based verification (optional, disabled by default)
    - Google ID token verification (optional, disabled by default)

    ## SMS Verification
    Uses Prelude for SMS verification. Each phone number:
    - Has a single pending verification at a time. Multiple `send_code` calls reuse the existing session.
    
    ## Lightning Network Verification
    Uses Phoenixd for Lightning Network invoice generation and payment tracking.
    The verification process involves:
    1. Creating a verification request to get a BOLT11 invoice
    2. Paying the invoice
    3. Checking payment status via the GET endpoint or polling the `/await` endpoint to wait for payment confirmation

    ## IP Verification
    A low-friction alternative to SMS/LN verification. A client POSTs to the endpoint and receives
    a signup code if their IP has not exceeded the configured weekly/annual rate limits.

    **Security note:** IP-based rate limiting is inherently weak — it relies on the reverse proxy
    setting `X-Forwarded-For` / `X-Real-IP` authoritatively. Users behind shared IPs (NAT, VPN)
...
    section from `config.toml`).

    ## Google Verification
...
    `config.toml`).
...
servers:
...
paths:
...
  /ln_verification:
    post:
      summary: Create Lightning Network verification
      description: |
        Creates a new Lightning Network payment verification request.
        Returns a BOLT11 invoice that must be paid to complete verification.
        
        The invoice expires after a configured time period (default 10 minutes).
        
        No request body is required.
      operationId: createLnVerification
      tags:
        - Lightning Network Verification
      responses:
        '200':
          description: Verification request created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateLnVerificationResponse'
              examples:
                success:
                  summary: Successful verification request
                  value:
                    id: "550e8400-e29b-41d4-a716-446655440000"
                    bolt11Invoice: "lnbc10u1p3qjqkapp5q..."
                    amountSat: 1000
                    expiresAt: 1704067200000
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
                example: "Internal Server Error"
...
  /ip_verification:
    post:
      summary: Create IP-based verification
      description: |
        Creates a new IP-based verification and returns a signup code.

        The client IP is determined from the socket address, or from `X-Forwarded-For` / `X-Real-IP`
        headers when `accept_proxy_ip_headers = true` in `config.toml`.

        Rate limits apply per IP address:
        - Weekly limit: 2 verifications per IP (default, configurable via `max_verifications_per_week` in `[ip_verification]`)
        - Annual limit: 4 verifications per IP (default, configurable via `max_verifications_per_year` in `[ip_verification]`)

        Signup tokens generated by this endpoint may include quota limits (storage, read/write rates)
        if an `[ip_verification.signup_quota]` section is present in `config.toml`. When configured,
        the homeserver POST endpoint is used to create tokens with those limits. When omitted,
        homeserver system defaults apply.

        This endpoint is only available when an `[ip_verification]` section is present in `config.toml`. Returns 404 otherwise.

        No request body is required.
      operationId: createIpVerification
      tags:
        - IP Verification
      responses:
        '200':
          description: Verification successful, signup code issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IpVerificationResponse'
              examples:
                success:
                  summary: Successful verification
                  value:
                    signupCode: "token-uuid-123"
                    homeserverPubky: "test-homeserver-pubky"
        '400':
          description: Could not determine client IP address
          content:
            text/plain:
              schema:
                type: string
                example: "Could not determine client IP address"
        '404':
          description: IP verification is not enabled on this server
        '429':
          description: Rate limit exceeded
          content:
            text/plain:
              schema:
                type: string
              examples:
                weeklyLimit:
                  value: "IP address has exceeded weekly verification limit"
                annualLimit:
                  value: "IP address has exceeded annual verification limit"
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
              examples:
                homeserverUnavailable:
                  value: "Homeserver temporarily unavailable, please retry"
                internalError:
                  value: "Internal Server Error"
  /google_verification:
    post:
      summary: Create Google ID token verification
      description: |
        Verifies a Google ID token server-side and returns a signup code if the token is valid
        and the Google identity has not exceeded the configured weekly/annual rate limits.

        The token must be an RS256-signed Google ID token for the configured `google_client_id`
        audience. The request body is limited to 20 KiB and unknown fields are rejected.

        Rate limits apply per verified Google identity (peppered hash of `iss` + `sub`):
        - Weekly limit: 2 verifications (default, configurable via `max_verifications_per_week` in `[google_verification]`)
        - Annual limit: 4 verifications (default, configurable via `max_verifications_per_year` in `[google_verification]`)

        This endpoint is only available when a `[google_verification]` section is present in `config.toml`. Returns 404 otherwise.
      operationId: createGoogleVerification
      tags:
        - Google Verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleVerificationRequest'
            examples:
              valid:
                summary: Valid Google ID token
                value:
                  googleIdToken: "eyJhbGciOiJSUzI1NiIsImtpZCI6Ii..."
      responses:
        '200':
          description: Verification successful, signup code issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoogleVerificationResponse'
              examples:
                success:
                  summary: Successful verification
                  value:
                    signupCode: "token-uuid-123"
                    homeserverPubky: "test-homeserver-pubky"
        '400':
          description: Malformed request (invalid JSON, unknown fields, empty token, or body over 20 KiB)
          content:
            text/plain:
              schema:
                type: string
                example: "invalid_request"
        '401':
          description: The Google ID token is invalid (bad signature, wrong audience/issuer, or expired)
          content:
            text/plain:
              schema:
                type: string
                example: "invalid_google_id_token"
        '404':
          description: Google verification is not enabled on this server
        '429':
          description: Rate limit exceeded for this Google identity
          content:
            text/plain:
              schema:
                type: string
              examples:
                weeklyLimit:
                  value: "weekly_limit_exceeded"
                annualLimit:
                  value: "annual_limit_exceeded"
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
              examples:
                homeserverUnavailable:
                  value: "homeserver_unavailable"
                googleVerifierUnavailable:
                  value: "google_verifier_unavailable"
                internalError:
                  value: "internal_error"
...
components:
  schemas:
...
    CreateLnVerificationResponse:
      type: object
      required:
        - id
        - bolt11Invoice
        - amountSat
        - expiresAt
      properties:
        id:
          $ref: '#/components/schemas/VerificationId'
          description: Verification ID for this verification request
        bolt11Invoice:
          type: string
          description: BOLT11 Lightning Network invoice string
          example: "lnbc10u1p3qjqkapp5q..."
        amountSat:
          type: integer
          format: int64
          minimum: 0
          description: Invoice amount in satoshis
          example: 1000
        expiresAt:
          type: integer
          format: int64
          description: Unix timestamp in milliseconds when the invoice expires
          example: 1704067200000
...
    IpVerificationResponse:
      type: object
      required:
        - signupCode
        - homeserverPubky
      properties:
        signupCode:
          type: string
          description: Signup token to use for homeserver registration
          example: "token-uuid-123"
        homeserverPubky:
          type: string
          description: Public key of the homeserver
          example: "test-homeserver-pubky"
    GoogleVerificationRequest:
      type: object
      required:
        - googleIdToken
      additionalProperties: false
      properties:
        googleIdToken:
          type: string
          description: RS256-signed Google ID token issued for the configured client ID
          example: "eyJhbGciOiJSUzI1NiIsImtpZCI6Ii..."
    GoogleVerificationResponse:
      type: object
      required:
        - signupCode
        - homeserverPubky
      properties:
        signupCode:
          type: string
          description: Signup token to use for homeserver registration
          example: "token-uuid-123"
        homeserverPubky:
          type: string
          description: Public key of the homeserver
          example: "test-homeserver-pubky"
...
    VerificationId:
      type: string
      format: uuid
      description: |
        Lightning Network verification ID.
        Must be a valid UUID (e.g., 550e8400-e29b-41d4-a716-446655440000).
        The value is normalized to lowercase.
      example: "550e8400-e29b-41d4-a716-446655440000"
...